#461 refactor on sys for dropping tables

+++ aaran/Testing/Database/Migrations/2024_06_24_000611_create_sw_tests_table.php
@@ -4,28 +4,30 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('sw_tests', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('module_id')->references('id')->on('test_modules')->onDelete('cascade');
-            $table->foreignId('admin_id')->references('id')->on('admin_tests')->onDelete('cascade');
-            $table->string('vname');
-            $table->longText('description');
-            $table->boolean('checked_1');
-            $table->boolean('checked_2');
-            $table->boolean('checked_3');
-            $table->boolean('checked_4');
-            $table->longText('comment');
-            $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->smallInteger('active_id');
-            $table->timestamps();
-        });
+        if (Aaran\Aadmin\Src\DbMigration::hasDeveloperTesting()) {
+
+            Schema::create('sw_tests', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('module_id')->references('id')->on('test_modules')->onDelete('cascade');
+                $table->foreignId('admin_id')->references('id')->on('admin_tests')->onDelete('cascade');
+                $table->string('vname');
+                $table->longText('description');
+                $table->boolean('checked_1');
+                $table->boolean('checked_2');
+                $table->boolean('checked_3');
+                $table->boolean('checked_4');
+                $table->longText('comment');
+                $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->smallInteger('active_id');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
